+++ src/Radicle/Internal.hs
@@ -5,12 +5,13 @@ module Radicle.Internal
 
 import           Protolude
 
+import           Radicle.Internal.CLI as X
 import           Radicle.Internal.Core as X
+import           Radicle.Internal.Interpret as X
 import           Radicle.Internal.Parse as X
 import           Radicle.Internal.Pretty as X
 import           Radicle.Internal.Primops as X
-import           Radicle.Internal.Subscriber as X
-import           Radicle.Internal.CLI as X
+import           Radicle.Internal.Effects as X
 
 import qualified Text.Megaparsec as M
