+++ src/Radicle/Internal/Parse.hs
@@ -9,7 +9,7 @@ import qualified Data.Text as T
 import           GHC.Exts (IsString(..))
 import           Text.Megaparsec (ParsecT, State(..), between, choice,
                                   defaultTabWidth, eof, initialPos, manyTill,
-                                  runParserT, runParserT', sepBy, try, (<?>))
+                                  runParserT', sepBy, try, (<?>))
 import qualified Text.Megaparsec as M
 import           Text.Megaparsec.Char (char, satisfy, space1)
 import qualified Text.Megaparsec.Char.Lexer as L
@@ -109,58 +109,6 @@ valueP = do
   pure v
 
 -- * Utilities
-
--- | Parse and evaluate a Text.
---
--- Examples:
---
--- >>> import Control.Monad.Identity
--- >>> import Radicle.Internal.Primops
--- >>> runIdentity $ interpret "test" "((lambda (x) x) #t)" pureEnv
--- Right (Boolean True)
---
--- >>> import Control.Monad.Identity
--- >>> runIdentity $ interpret "test" "(#t #f)" pureEnv
--- Left (TypeError "Trying to apply a non-function")
-interpret
-    :: Monad m
-    => Text       -- ^ Name of source file (for error reporting)
-    -> Text       -- ^ Source code to be interpreted
-    -> Bindings m -- ^ Bindings to be used
-    -> m (Either (LangError Value) Value)
-interpret sourceName expr bnds = do
-    let primopNames = Map.keys (bindingsPrimops bnds)
-        parsed = runReader (runParserT (spaceConsumer *> valueP <* eof) (toS sourceName) expr) primopNames
-    case parsed of
-        Left e  -> pure . Left $ ParseError e
-        Right v -> fst <$> runLang bnds (eval v)
-
--- | Parse and evaluate a Text as multiple expressions.
---
--- The first argument is the name of the source file to be used for error
--- reporting.
---
--- Examples:
---
--- >>> import Radicle.Internal.Primops
--- >>> fmap fst <$> runLang pureEnv $ interpretMany "test" "(define id (lambda (x) x))\n(id #t)"
--- Right (Boolean True)
-interpretMany
-    :: Monad m
-    => Text  -- ^ Name of source file (for error reporting)
-    -> Text  -- ^ Source code to be interpreted
-    -> Lang m Value
-interpretMany sourceName src = do
-    primopNames <- gets $ Map.keys . bindingsPrimops
-    let parsed = parseValues sourceName src primopNames
-    case partitionEithers parsed of
-        ([], vs) -> do
-          es <- mapM eval vs
-          case lastMay es of
-             Just e -> pure e
-             _ -> throwError
-                $ OtherError "InterpretMany should be called with at least one expression."
-        (e:_, _) -> throwError $ ParseError e
 
 -- | Parse a Text as a series of values.
 --
